move get_course_overview to __init__.py
